Makefile improvements, assembly listing file and print sizes
Renamed .obj to .elf as I found it confusing to have both .o and .obj
and I see it listed as .elf in another Makefile.

The extended assembly listing file includes the source code along
with the assembly which helps to identify what the compiler is
doing with that line of code.

Print the last and current section sizes.  It is useful to know how
big it is and how it is changing.

Ignore various output files.