Contest entry by David Fries, titled "capture"
Press a button or two to begin.  The lights start moving at the top
left or right.  If they are moving right you must hit the bottom right
button, opposite for left.  Two points if it is captured on 8, one
point if it is captured on 7 or 9, loose a try any other place, you
get three tries before the game starts over.  The score is displayed
as follows.  A sweep to 10 indicates the ten's digit is the next LED,
a sweep to 1 indicates a 1's digit, if either is 0 all LEDs stay dark.
The game gets faster the further along you get.

Subtitle First Time Quality.

The First Time Quality, because no kidding after the "EEMEM isn't
working, use PROGMEM instead" commit was just figuring out audio, all
the game logic was coded up with only one write to the AVR ATtiny
chip, and that is this version.

EEPROM failed me again, good enough so I'm going with it.