bug fix, getting a high score gets wraps turns around, fix score repeat
The math and logic runs fine in the software emulator for repeating
the score display.  There it would display 10's, 1's, then as data
counted up
led=1<<(12-data);
would continue to go negative causing two sweeps top and bottom,
pause, a number of times before data became 0 and repeated the
score.  On the hardware it would display 5 for a bit, pause, 1 for a
longer time, then nothing until it repeated.  Without looking I'm
guessing it has something to do with 16 bit unsigned subtraction using
8 bit registers.

DisplayScore advances to the next state after it is completed.  Except
the next state was COUNT_DOWN so after getting a high score tries
was still 0 and the next miss would decrement it to 255 and cause
it to take the long way around.